+++ src/Command/InstallCommand.php
@@ -2,29 +2,17 @@
 
 namespace Kapcus\DbChanger\Command;
 
-use Kapcus\DbChanger\Model\Exception\ConfigurationException;
 use Kapcus\DbChanger\Model\Exception\DbChangeException;
 use Kapcus\DbChanger\Model\Exception\EnvironmentException;
 use Kapcus\DbChanger\Model\Exception\InstallationException;
-use Kapcus\DbChanger\Model\Executor;
 use Kapcus\DbChanger\Model\IConfigurator;
-use Kapcus\DbChanger\Model\IExecutor;
-use Kapcus\DbChanger\Model\ILoader;
-use Kapcus\DbChanger\Model\DibiStorage;
 use Kapcus\DbChanger\Model\Manager;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class InstallCommand extends Command
+class InstallCommand extends FormattedOutputCommand
 {
-	/**
-	 * @var \Kapcus\DbChanger\Model\ILoader
-	 */
-	public $loader;
-
 	/**
 	 * @var \Kapcus\DbChanger\Model\Manager
 	 */
@@ -35,8 +23,7 @@ class InstallCommand extends Command
 	 */
 	public $configurator;
 
-	public function __construct(ILoader $loader, Manager $manager, IConfigurator $configurator) {
-		$this->loader = $loader;
+	public function __construct(Manager $manager, IConfigurator $configurator) {
 		$this->manager = $manager;
 		$this->configurator = $configurator;
 		parent::__construct();
@@ -49,7 +36,6 @@ class InstallCommand extends Command
 			->setDescription('Install registered DbChange on given environment.')
 			->addArgument('env', InputArgument::REQUIRED, 'Target environment code')
 			->addArgument('code', InputArgument::REQUIRED, 'DbChange code to be installed');
-			//->addOption('skip', 's', InputOption::VALUE_OPTIONAL, 'If manual parts of db change should be skipped or not.', true);
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output)