+++ src/Command/MarkCommand.php
@@ -2,42 +2,23 @@
 
 namespace Kapcus\DbChanger\Command;
 
+use Kapcus\DbChanger\Entity\Installation;
 use Kapcus\DbChanger\Entity\InstalledFragment;
 use Kapcus\DbChanger\Model\Exception\DbChangeException;
-use Kapcus\DbChanger\Model\Exception\EnvironmentException;
-use Kapcus\DbChanger\Model\IConfigurator;
-use Kapcus\DbChanger\Model\IGenerator;
-use Kapcus\DbChanger\Model\ILoader;
-use Kapcus\DbChanger\Model\DibiStorage;
 use Kapcus\DbChanger\Model\Manager;
-use Kapcus\DbChanger\Model\Util;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class MarkCommand extends Command
+class MarkCommand extends FormattedOutputCommand
 {
-	/**
-	 * @var \Kapcus\DbChanger\Model\ILoader
-	 */
-	public $loader;
-
 	/**
 	 * @var \Kapcus\DbChanger\Model\Manager
 	 */
 	public $manager;
 
-	/**
-	 * @var \Kapcus\DbChanger\Model\IConfigurator
-	 */
-	public $configurator;
-
-	public function __construct(ILoader $loader, Manager $manager, IConfigurator $configurator) {
-		$this->loader = $loader;
+	public function __construct(Manager $manager) {
 		$this->manager = $manager;
-		$this->configurator = $configurator;
 		parent::__construct();
 	}
 
@@ -45,23 +26,19 @@ class MarkCommand extends Command
 	{
 		$this
 			->setName('dbchanger:mark')
-			->setDescription(sprintf('Mark DbChange fragment with given status: %s.', InstalledFragment::getStatusNameString()))
-			->addArgument('fragmentId', InputArgument::REQUIRED, 'Fragment id.')
+			->setDescription(sprintf('Mark DbChange fragment(s)/installation with given status: %s / %s.', InstalledFragment::getStatusNameString(), Installation::getStatusNameString()))
+			->addArgument('target', InputArgument::REQUIRED, 'Fragment id, fragment range or installation id.')
 		->addArgument('status', InputArgument::REQUIRED, 'Status to be set.');
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output)
 	{
-		$fragmentId = strtoupper($input->getArgument('fragmentId'));
+		$target = strtoupper($input->getArgument('target'));
 		$statusShortcut = strtoupper($input->getArgument('status'));
 
 		try {
-			if (Util::isFragmentId($fragmentId)) {
-				$this->manager->markFragmentById(Util::getIdFromFragmentId($fragmentId), $statusShortcut);
-			} else {
-				throw new DbChangeException(sprintf('Invalid fragment id %s.', $fragmentId));
-			}
-			$output->writeln(sprintf('Fragment %s successfully marked with status %s.', $fragmentId, InstalledFragment::getStatusName(InstalledFragment::getStatusByShortcut($statusShortcut))));
+			$this->manager->markTarget($target, $statusShortcut);
+			$output->writeln(sprintf('Target successfully marked with status %s.', InstalledFragment::getStatusName(InstalledFragment::getStatusByShortcut($statusShortcut))));
 		} catch (DbChangeException $e) {
 			$output->writeln($e->getMessage());
 		}