+++ src/Command/RegisterCommand.php
@@ -3,19 +3,14 @@
 namespace Kapcus\DbChanger\Command;
 
 use Kapcus\DbChanger\Model\Exception\DbChangeException;
-use Kapcus\DbChanger\Model\Exception\EnvironmentException;
-use Kapcus\DbChanger\Model\IConfigurator;
-use Kapcus\DbChanger\Model\IGenerator;
 use Kapcus\DbChanger\Model\ILoader;
-use Kapcus\DbChanger\Model\DibiStorage;
 use Kapcus\DbChanger\Model\Manager;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class RegisterCommand extends Command
+class RegisterCommand extends FormattedOutputCommand
 {
 	/**
 	 * @var \Kapcus\DbChanger\Model\ILoader
@@ -27,15 +22,9 @@ class RegisterCommand extends Command
 	 */
 	public $manager;
 
-	/**
-	 * @var \Kapcus\DbChanger\Model\IConfigurator
-	 */
-	public $configurator;
-
-	public function __construct(ILoader $loader, Manager $manager, IConfigurator $configurator) {
+	public function __construct(ILoader $loader, Manager $manager) {
 		$this->loader = $loader;
 		$this->manager = $manager;
-		$this->configurator = $configurator;
 		parent::__construct();
 	}