+++ src/Command/ReinitCommand.php
@@ -2,16 +2,11 @@
 
 namespace Kapcus\DbChanger\Command;
 
-use Kapcus\DbChanger\Model\Exception\OutOfSyncException;
-use Kapcus\DbChanger\Model\IConfigurator;
-use Kapcus\DbChanger\Model\ILoader;
-use Kapcus\DbChanger\Model\Manager;
 use Nette\NotImplementedException;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class ReinitCommand extends Command
+class ReinitCommand extends FormattedOutputCommand
 {
 
 	public function __construct() {